# Conflicts:
#	core/pieShareApp/src/main/java/org/pieShare/pieShareApp/model/entities/PieFolderEntity.java
#	core/pieShareApp/src/main/java/org/pieShare/pieShareApp/model/entities/api/IPieFolderEntity.java
#	core/pieShareApp/src/main/java/org/pieShare/pieShareApp/model/pieFilder/PieFilder.java
#	core/pieShareApp/src/main/java/org/pieShare/pieShareApp/model/pieFilder/PieFolder.java
#	core/pieShareApp/src/main/java/org/pieShare/pieShareApp/service/database/DAOs/PieFileDAO.java
#	core/pieShareApp/src/main/java/org/pieShare/pieShareApp/service/database/DAOs/PieFolderDAO.java
#	core/pieShareApp/src/main/java/org/pieShare/pieShareApp/service/database/ModelEntityConverterService.java
#	core/pieShareApp/src/main/java/org/pieShare/pieShareApp/service/fileService/HistoryFileService.java
#	core/pieShareApp/src/main/java/org/pieShare/pieShareApp/service/folderService/FolderService.java
#	core/pieShareApp/src/main/java/org/pieShare/pieShareApp/service/folderService/IFolderService.java
#	core/pieShareApp/src/main/java/org/pieShare/pieShareApp/service/historyService/HistoryService.java
#	core/pieShareApp/src/main/java/org/pieShare/pieShareApp/task/localTasks/folderEventTask/ALocalFolderEventTask.java
#	core/pieShareAppIntegrationTests/src/test/java/pieShareAppITs/DataBaseIT.java

+++ piePlate/src/main/java/org/pieShare/pieTools/piePlate/service/cluster/zeroMqCluster/socket/PieRouter.java
@@ -8,10 +8,13 @@ package org.pieShare.pieTools.piePlate.service.cluster.zeroMqCluster.socket;
 import java.net.InetAddress;
 import java.nio.channels.ClosedSelectorException;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import javax.inject.Provider;
+import org.pieShare.pieTools.piePlate.model.DiscoveredMember;
 import org.pieShare.pieTools.piePlate.model.PieAddress;
 import org.pieShare.pieTools.piePlate.service.channel.api.IIncomingChannel;
+import org.pieShare.pieTools.piePlate.service.cluster.zeroMqCluster.IEndpointCallback;
 import org.pieShare.pieTools.piePlate.service.cluster.zeroMqCluster.socket.api.IPieRouter;
 import org.pieShare.pieTools.piePlate.task.ChannelTask;
 import org.pieShare.pieTools.pieUtilities.service.pieExecutorService.api.IExecutorService;
@@ -26,13 +29,12 @@ import org.zeromq.ZMQException;
  */
 public class PieRouter implements IPieRouter {
 
-	private ZContext context;
-	private ZMQ.Socket router;
 	private ZeroMQUtilsService utils;
 	private IExecutorService executorService;
 	private Provider<ChannelTask> channelTaskProvider;
 	private List<IIncomingChannel> incomingChannels;
 	private boolean shutdown;
+	private int port;
 
 	public PieRouter() {
 		this.incomingChannels = new ArrayList<>();
@@ -52,12 +54,31 @@ public class PieRouter implements IPieRouter {
 	}
 
 	@Override
-    public boolean bind(int port) {
-		if (context == null) {
-			this.context = new ZContext(1);
-			this.router = context.createSocket(ZMQ.ROUTER);
-		}
-		
+	public void setPort(int port) {
+		this.port = port;
+	}
+
+	@Override
+	public void close() {
+		shutdown = true;
+		PieDealer dealer = new PieDealer();
+		dealer.setZeroMQUtilsService(utils);
+		List<DiscoveredMember> members = new ArrayList<>();
+		DiscoveredMember member = new DiscoveredMember();
+		member.setInetAdresses(InetAddress.getLoopbackAddress());
+		member.setPort(port);
+		member.setName("mySelf");
+		members.add(member);
+		dealer.send(members, "!shutdown".getBytes(), new IEndpointCallback() {
+			@Override
+			public void nonRespondingEndpoint(List<DiscoveredMember> brokenMembers) {
+			}
+		});
+		dealer.shutdown();
+	}
+
+	@Override
+	public void run() {
 		//When not bound to 0.0.0.0 MAC will not see the port as bound but only
 		//in combination with the specific address which leads to wrong behaviour.
 		//More clear explanation:
@@ -72,58 +93,56 @@ public class PieRouter implements IPieRouter {
 		//router.bind(utils.buildConnectionString(address, port));
 		//see also NetworkService
 		String connectionString = utils.buildConnectionString("0.0.0.0", port);
-        
-        PieLogger.trace(PieRouter.class, "Bind router on: {}", connectionString);
+
+		ZContext context = new ZContext(1);
+		ZMQ.Socket router = context.createSocket(ZMQ.ROUTER);
+
+		PieLogger.trace(PieRouter.class, "Bind router on: {}", connectionString);
 		try {
 			router.bind(connectionString);
-			return true;
+
+			//clean shutdown, ZMQ poller?
+			//zeromq can't block and sleep
+			//https://github.com/zeromq/jeromq/blob/master/src/main/java/zmq/SocketBase.java#L712
+			//https://github.com/thriftzmq/thriftzmq-java/blob/master/thriftzmq/src/main/java/org/thriftzmq/ProxyLoop.java
+			byte[] msg = null;
+			while (!shutdown) {
+				try {
+					//receive dealer identity and discard it
+					router.recv();
+					msg = router.recv();
+				} catch (ZMQException | AssertionError | ClosedSelectorException e) {
+					//todo: can this really only happen when shutding down or 
+					//also in other scenarios?
+					PieLogger.warn(this.getClass(), "Exception in PieRouter! Shuting router down!", e);
+					return;
+				}
+				
+				//we need to shutdown in this strange way because ZMQ is not thread save
+				//we can not shutdown the socket prperly from another thread
+				
+				//todo: consider using a 2nd router for receiving the shutdown signal
+				//this way we would be independet of the incoming msgs
+				//however we would need another pattern to implement this
+				if (Arrays.equals(msg, "!shutdown".getBytes())) {
+					router.unbind(this.utils.buildConnectionString("0.0.0.0", port));
+					context.destroySocket(router);
+					context.destroy();
+					break;
+				}
+
+				for (IIncomingChannel channel : incomingChannels) {
+					ChannelTask task = this.channelTaskProvider.get();
+					task.setChannel(channel);
+					task.setMessage(msg);
+					PieAddress addr = new PieAddress();
+					task.setAddress(addr);
+
+					this.executorService.execute(task);
+				}
+			}
 		} catch (ZMQException e) {
 			PieLogger.error(PieRouter.class, "Bind failed: {}", e);
-			return false;
-		}
-	}
-
-	@Override
-	public void close() {
-		shutdown = true;
-
-		context.destroy();
-	}
-
-	@Override
-	public byte[] receive() {
-		//receive dealer identity and discard it
-		router.recv();
-
-		return router.recv();
-	}
-
-	@Override
-	public void run() {
-		byte[] msg = null;
-		//clean shutdown, ZMQ poller?
-		//zeromq can't block and sleep
-		//https://github.com/zeromq/jeromq/blob/master/src/main/java/zmq/SocketBase.java#L712
-		//https://github.com/thriftzmq/thriftzmq-java/blob/master/thriftzmq/src/main/java/org/thriftzmq/ProxyLoop.java
-		while (!shutdown) {
-			try {
-				msg = this.receive();
-			} catch (ZMQException | AssertionError | ClosedSelectorException e) {
-				//todo: can this really only happen when shutding down or 
-				//also in other scenarios?
-				PieLogger.warn(this.getClass(), "Exception in PieRouter! Shuting router down!", e);
-				return;
-			}
-			
-			for (IIncomingChannel channel : incomingChannels) {
-				ChannelTask task = this.channelTaskProvider.get();
-				task.setChannel(channel);
-				task.setMessage(msg);
-				PieAddress addr = new PieAddress();
-				task.setAddress(addr);
-
-				this.executorService.execute(task);
-			}
 		}
 
 	}
